requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2.1
django-ckeditor==6.5.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-js-asset==2.0.0
importlib-metadata==6.6.0
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0